backend/requirements.txt
fastapi
uvicorn[standard]
xgboost
scikit-learn
pandas
numpy
requests
python-dotenv
